requirements.txt
CacheControl==0.12.14
certifi==2024.2.2
charset-normalizer==3.3.2
deprecation==2.1.0
enum34==1.1.10
Faker==24.4.0
idna==3.6
jsonpickle==3.0.3
jsonpointer==2.4
msgpack==1.0.8
packaging==24.0
prettytable==3.10.0
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.31.0
setuptools==69.2.0
six==1.16.0
squareup
urllib3==2.2.1
wcwidth==0.2.13
